refactor: Comment 리팩토링 (#74)
