+++ src/features/board/components/Comment/styles/CommentItem.styled.ts
